feat: implement saveMessage in BulkSaveUserReviewsSlashCommandHnadler
